Improve flatten command with temp folder

+++ src/modules/FolderOperations.ts
@@ -11,32 +11,41 @@ const FolderOperations: RawModule = {
     run: async (opts: {all: boolean, filesAllowed: boolean}) => {
       console.log('Start flattening...');
 
-      await doForEach(ENV.cwd, (dir, folder) => {
-        if (!dir.isDirectory()) return;
+      const tmpPath = path.join(ENV.cwd, '__tmp__');
+      if (fs.existsSync(tmpPath))
+        console.warn(highlightExps`Caution: ${tmpPath} already exists, it will be emptied in ${ENV.cwd} after flattening.`);
+      else
+        fs.mkdirSync(tmpPath);
 
-        const currentFolder = path.join(folder, dir.name);
+      await doForEach(ENV.cwd, (dir, _) => {
+        if (!dir.isDirectory() || dir.path === tmpPath) return;
+
+        const currentFolder = dir.path;
         const ents = getEnts(currentFolder);
         if (ents.length === 0)
           return;
 
         if (!opts.filesAllowed && ents.some(ent => ent.isFile()))
           return;
-        
-        if (opts.all)
-          ents.forEach(nestedEnt => simpleMove(currentFolder, nestedEnt.name, folder));
-        else {
-          if (ents.length !== 1)
-            return;
-          simpleMove(currentFolder, ents[0].name, folder);
-        }
 
         try {
+          if (opts.all)
+            ents.forEach(nestedEnt => simpleMove(currentFolder, nestedEnt.name, tmpPath));
+          else {
+            if (ents.length !== 1)
+              return;
+            simpleMove(currentFolder, ents[0].name, tmpPath);
+          }
+
           fs.rmdirSync(currentFolder);
           console.log(highlightExps`Flattened ${dir.name} -> ${ents.map(e => `"${e.name}"`).join(', ')}`);
         } catch (err) {
           console.error(highlightExps`Flattening ${dir.name} failed:`, err);
         }
       });
+
+      getEnts(tmpPath).forEach(e => simpleMove(tmpPath, e.name, ENV.cwd));
+      fs.rmdirSync(tmpPath);
     }
   },
   clean: {
